requirements.txt
pandas
numpy
python-slugify
dill
sklearn
sklearn_pandas
sklearn_porter
imblearn
